Testing operations (Patient, Medecin, Consultation, RendezVous)

+++ src/main/java/com/jpa/orm_data/repositories/RendezVousRepository.java
@@ -3,5 +3,5 @@ package com.jpa.orm_data.repositories;
 import com.jpa.orm_data.entities.RendezVous;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RendezVousRepository extends JpaRepository<RendezVous, Integer> {
+public interface RendezVousRepository extends JpaRepository<RendezVous, Long> {
 }
